src/main/java/com/sergey/zelenov/shiftbooking/service/ShiftService.java
package com.sergey.zelenov.shiftbooking.service;

import com.sergey.zelenov.shiftbooking.model.Shift;
import com.sergey.zelenov.shiftbooking.repository.ShiftRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ShiftService {

    @Autowired
    private ShiftRepository shiftRepository;

    public Shift createShift(Shift shift) {
        return shiftRepository.save(shift);
    }

    public Shift updateShift(Long shiftId, Shift updatedShift) {
        Shift shift = shiftRepository.findById(shiftId)
                .orElseThrow(() -> new RuntimeException("Shift not found"));
        shift.setTitle(updatedShift.getTitle());
        shift.setDescription(updatedShift.getDescription());
        shift.setLocation(updatedShift.getLocation());
        shift.setHourlyRate(updatedShift.getHourlyRate());
        shift.setStartTime(updatedShift.getStartTime());
        shift.setEndTime(updatedShift.getEndTime());
        return shiftRepository.save(shift);
    }

    public void deleteShift(Long shiftId) {
        shiftRepository.deleteById(shiftId);
    }

    public List<Shift> getAllShifts() {
        return shiftRepository.findAll();
    }

    public Shift getShiftById(Long shiftId) {
        return shiftRepository.findById(shiftId)
                .orElseThrow(() -> new RuntimeException("Shift not found"));
    }
}
